moved types into higher scope
