+++ src/server/src/types/embedding.ts
@@ -1,0 +1,4 @@
+export interface EmbeddingVector {
+    embedding: number[];
+    chunk_text: string;
+}
