+++ src/server/src/types/scraping.ts
@@ -1,0 +1,7 @@
+export interface ScrapingResponse{
+	title: string, 
+	episodeTitle: string,
+	season: number,
+	episode: number,
+	transcript: string
+}
