+++ src/server/src/utils/scraper.ts
@@ -1,13 +1,7 @@
 import axios from "axios";
 import * as cheerio from "cheerio";
+import { ScrapingResponse } from "types/scraping";
 
-export interface ScrapingResponse{
-	title: string, 
-	episodeTitle: string,
-	season: number,
-	episode: number,
-	transcript: string
-}
 
 const BASE_URL = "https://transcripts.foreverdreaming.org";
 
@@ -38,7 +32,6 @@ export async function scrapeTranscript(topicId: string): Promise<ScrapingRespons
 	try {
 		const link = `${BASE_URL}/viewtopic.php?t=${topicId}`;
 		
-		// Add delay to avoid rate limiting
 		await new Promise(resolve => setTimeout(resolve, 1000));
 		
 		const { data } = await axios.get(link, {
@@ -84,35 +77,4 @@ export async function scrapeTranscript(topicId: string): Promise<ScrapingRespons
 			throw error;
 		}
 	}
-}
-
-
-// export async function scrapeAllTranscripts(pageExtension: number) {
-// 	const { showName, episodes } = await scrapeOverview(pageExtension);
-// 	console.log(`Found ${episodes.length} episodes for ${showName}. Scraping transcripts...`);
-
-// 	const results = [];
-// 	for (const ep of episodes) {
-// 		if (!ep.topicId) continue; 
-// 		const transcriptData = await scrapeTranscript(ep.topicId);
-// 		results.push({ ...ep, ...transcriptData });
-// 	}
-
-// 	return { showName, episodes: results };
-// }
-
-if (require.main === module) {
-	(async () => {
-		const [, , type, id] = process.argv;
-		if (type === "overview") {
-			console.log(await scrapeOverview(Number(id)));
-		} else if (type === "transcript") {
-			const response = await scrapeTranscript(id);
-			console.log(response);
-		} else if (type === "all") {
-			// console.log(await scrapeAllTranscripts(Number(id)));
-		} else {
-			console.log("Usage: npx ts-node scraper.ts <overview|transcript|all> <pageExtensionOrTopicId>");
-		}
-	})();
 }
